refactor: RandomGenerator.generate()를 테스트가능한 구조로 수정
