--- src/main/java/racing/RandomGenerator.java
@@ -1,11 +1,0 @@
-package racing;
-
-import java.util.Random;
-
-public class RandomGenerator {
-  private static final Random random = new Random();
-
-  public static int generate() {
-    return random.nextInt(10);
-  }
-}
